requirements.txt
wheezy.template
flask
requests
SPARQLWrapper==1.8.5
flasgger==0.9.1